Fix bug in gate target

+++ train.py
@@ -104,10 +104,9 @@ def main(args):
                 gate_target = mel_pos.eq(0).float()
                 print(gate_target)
                 index_arr = torch.Tensor(
-                    [i for i in range(mel_length.size(0))]).int().to(device)
+                    [i for i in range(mel_length.size(0))]).long().to(device)
                 index_arr = torch.cat(
-                    [index_arr.unsqueeze(1), (mel_length-1).int().unsqueeze(1)], 1)
-                index_arr = index_arr.int()
+                    [index_arr.unsqueeze(1), (mel_length-1).unsqueeze(1)], 1)
                 # print(index_arr)
                 gate_target[index_arr] = 1.
                 print(gate_target)
